+++ tests/V5/Test/Highlights.hs
@@ -7,28 +7,24 @@ import Test.Import
 
 import qualified Data.Map as M
 
+initHighlights :: Text -> BH IO (Either EsError (Maybe HitHighlight))
+initHighlights fieldName = do
+  _ <- insertData
+  _ <- insertOther
+  let query = QueryMatchQuery $ mkMatchQuery (FieldName fieldName) (QueryString "haskell")
+  let testHighlight = Highlights Nothing [FieldHighlight (FieldName fieldName) Nothing]
+  let search = mkHighlightSearch (Just query) testHighlight
+  searchTweetHighlight search
+
 spec :: Spec
 spec =
   describe "Highlights API" $ do
-
     it "returns highlight from query when there should be one" $ withTestEnv $ do
-      _ <- insertData
-      _ <- insertOther
-      let query = QueryMatchQuery $ mkMatchQuery (FieldName "message") (QueryString "haskell")
-      let testHighlight = Highlights Nothing [FieldHighlight (FieldName "message") Nothing]
-
-      let search = mkHighlightSearch (Just query) testHighlight
-      myHighlight <- searchTweetHighlight search
+      myHighlight <- initHighlights "message"
       liftIO $
         myHighlight `shouldBe` Right (Just (M.fromList [("message",["Use <em>haskell</em>!"])]))
 
     it "doesn't return highlight from a query when it shouldn't" $ withTestEnv $ do
-      _ <- insertData
-      _ <- insertOther
-      let query = QueryMatchQuery $ mkMatchQuery (FieldName "message") (QueryString "haskell")
-      let testHighlight = Highlights Nothing [FieldHighlight (FieldName "user") Nothing]
-
-      let search = mkHighlightSearch (Just query) testHighlight
-      myHighlight <- searchTweetHighlight search
+      myHighlight <- initHighlights "user"
       liftIO $
         myHighlight `shouldBe` Right Nothing
